Fix module rebuilding and stdlib version matching

+++ start.js
@@ -35,10 +35,21 @@ if(fs.existsSync('./update.7z')) {
         process.exit(-1);
     }
 
-    fs.rmdirSync('./bin', {recursive: true});
-    fs.rmdirSync('./node_modules', {recursive: true});
+    if(fs.existsSync('./bin')) {
+        fs.rmdirSync('./bin', {recursive: true});
+    }
+
+    if(fs.existsSync('./node_modules')) {
+        fs.rmdirSync('./node_modules', {recursive: true});
+    }
+
+    if(fs.existsSync('./modules/tswow-stdlib')) {
+        fs.rmdirSync('./modules/tswow-stdlib',{recursive: true});
+    }
+
     fs.renameSync('./tmp/bin','./bin')
     fs.renameSync('./tmp/node_modules','./node_modules');
+    fs.renameSync('./tmp/modules/tswow-stdlib','./modules/tswow-stdlib');
 
     fs.rmdirSync('./tmp', {recursive: true});
 }
